requirements.txt
Django==6.0.4
djangorestframework==3.17.1
psycopg[binary]==3.3.3
channels==4.3.1
daphne==4.2.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.6.0
